feat: Refactor comment reply logic and improve error handling
- Updated NewGetCommentRepliesLogic function documentation for clarity.
- Simplified parameter handling in GetCommentReplies method by removing unnecessary type conversions.
- Enhanced error handling in CreateComment method by ignoring the result of BatchDeleteRelation to prevent potential issues.
- Improved UploadAvatar method to include additional parameters (Filename, ContentType) for better image upload handling.

These changes enhance the clarity and efficiency of comment and avatar management functionalities.

+++ api/TechSphere/Comment/internal/logic/getcommentreplieslogic.go
@@ -16,7 +16,7 @@ type GetCommentRepliesLogic struct {
 	svcCtx *svc.ServiceContext
 }
 
-// 获取某条评论的子评论列表
+// NewGetCommentRepliesLogic 获取某条评论的子评论列表
 func NewGetCommentRepliesLogic(ctx context.Context, svcCtx *svc.ServiceContext) *GetCommentRepliesLogic {
 	return &GetCommentRepliesLogic{
 		Logger: logx.WithContext(ctx),
@@ -28,8 +28,8 @@ func NewGetCommentRepliesLogic(ctx context.Context, svcCtx *svc.ServiceContext) 
 func (l *GetCommentRepliesLogic) GetCommentReplies(req *types.GetCommentRepliesReq) (resp *types.GetCommentRepliesResp, err error) {
 	reply, err := l.svcCtx.CommentService.GetCommentReplies(l.ctx, &comment.GetCommentRepliesRequest{
 		CommentId: req.CommentId,
-		Page:      int32(req.Page),
-		PageSize:  int32(req.PageSize),
+		Page:      req.Page,
+		PageSize:  req.PageSize,
 	})
 	if err != nil {
 		return nil, err
